pyproject.toml
[tool.poetry]
name = "chess_ai"
version = "0.1.0"
description = "Chess Game with Neural Network AI"
authors = ["Peyton Anderson <peyanders@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
pygame = "^2.3.0"
pyyaml = "^6.0"

[tool.poetry.dev-dependencies]
